TS-2: OSX port, this checkin has traffic_server running on OSX (thanx to georgep) TS-52: this checkin encapsulates the networking code and enables libev support which was used for the OSX port (TS-2 above) (jplevyak)
git-svn-id: https://svn.apache.org/repos/asf/incubator/trafficserver/traffic/branches/dev@892325 13f79535-47bb-0310-9956-ffa450edef68

+++ proxy/Update.cc
@@ -2323,7 +2323,7 @@ HtmlParser::ValidSupportedProtoScheme(char *p)
 int
 HtmlParser::ExtractURL(char **url, char **url_end)
 {
-  int n;
+  intptr_t n;
 
   // '#' considerations
   if (_attr_value_hash_char_index >= 0) {
@@ -2341,7 +2341,7 @@ HtmlParser::ExtractURL(char **url, char **url_end)
     if (_html_doc_base) {
       _html_doc_base.clear();
     }
-    int n;
+    intptr_t n;
     for (n = 0; n < _attr_value.length(); ++n) {
       _html_doc_base(_html_doc_base.length()) = _attr_value[n];
     }
@@ -2388,7 +2388,7 @@ HtmlParser::ExtractURL(char **url, char **url_end)
   }
 
   if (_attr_value.length() > 1) {
-    *url = &_attr_value[0];
+    *url = &_attr_value[(intptr_t)0];
     *url_end = &_attr_value[_attr_value.length() - 2];
     return 1;
 
@@ -2474,7 +2474,7 @@ HtmlParser::ConstructURL(char **url, char **url_end)
     if (base)
       delete base;
 
-    *url = &_result[0];
+    *url = &_result[(intptr_t)0];
     *url_end = &_result[_result.length() - 3];  // -1 (real len)
     // -1 (skip null)
     // -1 (zero base)
