+++ proxy/mgmt2/cli/CLIlineBuffer.cc
@@ -97,7 +97,7 @@ CLIlineBuffer::numFields()
 int
 CLIlineBuffer::getLineSize()
 {
-  int i;
+  intptr_t i;
   int line_size = 0;
 
   for (i = 0; i < c_fields; i++)
@@ -110,7 +110,7 @@ CLIlineBuffer::getLineSize()
 int
 CLIlineBuffer::getStringSize()
 {
-  int i;
+  intptr_t i;
   int s_size = 0;
 
   for (i = 0; i < c_fields; i++)
@@ -130,7 +130,7 @@ CLIlineBuffer::getStringSize()
 int
 CLIlineBuffer::getDepth()
 {
-  int i;
+  intptr_t i;
   int cdepth = 1;
   int depth = 1;
   int spdepth = 0;
@@ -212,7 +212,7 @@ CLIlineBuffer::getline()
   int iters = 0;
   static char *cptr = '\0';
   bool done = false;
-  int i;
+  intptr_t i;
 
   // calculate output line buffer size
   buf_size = buf_depth * (line_size + strlen("\n") + 1) + 1;
